+++ CHANGELOG.md
@@ -10,6 +10,7 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ### Fixed
 
 * Moved auth query parameter value for trace/log/replay exporters into X-SF-Token header.
+* 'deploymentEnvironment', in configuration, must now contain a non-empty string.
 
 ## [2.0.4] - 2025-12-01